Fixed JFXC-3856 : TextBox shows no cursor.
In decompose, the existing code avoided shreding of "defs". Fix is to
check either if var is not def or if it is def and bound. This way bound def
variable cases are shreded properly.

Thanks to Jim Holmlund for the sdk independent simple test case -- compiling
that test with -XDdumpdepgraph=. option immediately showed the dependency
problem! (wish for such nice simple test cases for all the bugs ;-) )

One minor issue with -XDdumpdepgraph. For overridden variables, the graph
shows SuperClass.varName instead of SubClass.varName. (yet to fix this).

Testing:

* added a regression test from JIRA report.
* ran "ant dev-pass" on Mac OS X 10.5.8.
* built compiler-scrum with the fixed compiler and tried one-liner test

    var testArea = javafx.scene.control.TextBox{text: "kick me"};

Manually checked that the text box shows cursor when focus is on text box.