Progress: JFXC-3917
All the nodes and information required to refactor the on replace claues in JavaFX was in fact already there. However the JavaFXTreeScanner did not contain code that shows how this should be done. Hence this change updates visitOnReplace to show that the start and en indexe variables are avaiable in the node.