Fixed: JFXC-3717
The tree node SequenceInsert now allows the porgrammer to call the method getPosition() and check for null. If null then there was no insert position. If there was an insert position, then call shouldInsertAfter() [which is already public] to find out if this is before/after. You get the start and endpos of the insert index expression by looking at the expression returned from getPosition() and asking for the start and endpos in whichever way is most convenient (I will look to see if JavafxTreeInfo supports JFXExpression).