Fixed: JFXC-4044
The tree maker code for TimeLiteral erroneously used a Double for the literal only if it found a '.' in the literal specification. This obviously tried to install as integer time literals, values that looked like integer values but were actually too big to be held in Integers. This code is now fixed to produce a Double version of the literal and only utliize an Integer if the value is within the Integer MIN_VAL/MAX_VAL range and is an integer quantity.