Remove array-length check in generated invoke$ functions.
This is related to JFXC-3530 "Too much bytecode is generated for =>"
and the function-value changes JimL checked in (unknown JIRAs).

Calling an unknown function calls an invoke$ method in a function's
handler; the invoke$ takes an array of argument values, and does a
switch on a function number.  Each case checks that the array has the
expected length.  This check is redundant, since static type-checking
will verify that we're passing the correct number of arguments.  Thus
we're needlessly adding static footprint, and at run-time executing
needless code.

(It's probably possible to do Java-level casts in some way
to end up with the wrong type, but guarding against this
possibility does not seem worth the overhead.)

Reviewed by Jim Laskey.