Fix JFXC-4170 - avoid allocating array for functon value arguments.
Invokeing a function is changed to use this general-purpose method:
    public Object invoke$(Object arg1, Object arg2, Object[] rargs)
If we have 0 arguments, then arg1, arg2 and rargs are null.  If we
have 1 arguments, then it is passed in arg1, while arg2 and rargs are
null.  If we have 2 arguments, they are passed in arg1 and arg2, while
rargs is null.  If we have more than 2 arguments, the first 2 arg
passed in arg1 and arg2, while the rest are passed in rargs.