Fixed: JFXC-4087
Because the api calls must duplicate portions of a tree in order to return them to the caller of getScope(), the endpos map was out of sync with the objects that represent the tree. The JavafxTreeCopier now adds newly duplicated nodes to the endpos map so that both the old and new objects are represented.

Added test case to test that the endpos positions are valid.
Successfully ran ant dev-pass