Fixed: JFXC-3614
The script class builder woudl always reorder the run method so that it came first in the tree. This was just a side effect of it culling out a user defined run method though and was not necessary. This change causes JFXTree to implement the Comparible interface and for the class builder to sort the list of class members by 'pos' order after picking out the run method.

Junit test supplied
ant dev-pass completed successfully