Fixed: JFXC-2751
Because the langauge is syntactically ambiguous we cannot detect illegal attempts to call something that are caused because of a missing separator on a definition.
Hence I have improved the error message that comes from attribution when it determines that a type is not somethign that is callable but someone
seems to be trying to call it anyway.

ant dev-pass ran successfully.